Fix Controller
