--- HW7/Seva_Popov/TradingSimulatorWebApi/TradingSimulator.Core/Repositories/IUserAndSharesTableRepository.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace TradingSimulator.Core.Repositories
-{
-    public interface IUserAndSharesTableRepository
-    {
-    }
-}
