--- HW7/Seva_Popov/TradingSimulatorWebApi/TradingSimulatorConsoleApp/Services/UserAndSharesService.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace TradingSimulatorConsoleApp.Services
-{
-    class UserAndSharesService
-    {
-    }
-}
