--- HW7/Seva_Popov/TradingSimulatorWebApi/TradingSimulatorWebApi/Controllers/ValuesController.cs
@@ -1,50 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using TradingSimulator.Core.Models;
-using TradingSimulatorConsoleApp.Data;
-
-
-namespace TradingSimulatorWebApi.Controllers
-{
-    [Route("api/[controller]")]
-    [ApiController]
-    public class ValuesController : ControllerBase
-    {
-        TradingSimulatorDbContext db = new TradingSimulatorDbContext();
-
-        // GET api/values
-        [HttpGet]
-        public IEnumerable<UserEntity> Get()
-        {
-            return db.Users.ToList();
-        }
-
-        // GET api/values/5
-        [HttpGet("{id}")]
-        public ActionResult<string> Get(int id)
-        {
-            return "value";
-        }
-
-        // POST api/values
-        [HttpPost]
-        public void Post([FromBody] string value)
-        {
-        }
-
-        // PUT api/values/5
-        [HttpPut("{id}")]
-        public void Put(int id, [FromBody] string value)
-        {
-        }
-
-        // DELETE api/values/5
-        [HttpDelete("{id}")]
-        public void Delete(int id)
-        {
-        }
-    }
-}
